+++ src/harmony/components/tetradic.ts
@@ -4,13 +4,12 @@ import { HSL } from '../../types';
 /**
  * Generates a tetradic (rectangle) color harmony.
  * @param {Color} color - The base color.
- * @param {number} [angle=60] - The angle between color pairs (default: 60 degrees).
  * @returns {[Color, Color, Color, Color]} An array containing the base color and three tetradic colors.
  */
-export function tetradic(color: Color, angle: number = 60): [Color, Color, Color, Color] {
+export function tetradic(color: Color): [Color, Color, Color, Color] {
     const hsl = color.toHsl();
     const color1: HSL = {
-        h: (hsl.h + angle) % 360,
+        h: (hsl.h + 90) % 360,
         s: hsl.s,
         l: hsl.l
     };
@@ -20,7 +19,7 @@ export function tetradic(color: Color, angle: number = 60): [Color, Color, Color
         l: hsl.l
     };
     const color3: HSL = {
-        h: (hsl.h + 180 + angle) % 360,
+        h: (hsl.h + 270) % 360,
         s: hsl.s,
         l: hsl.l
     };